feat: implement basic supabase ssr
